update JS SDK guide regarding JWT issuer
As of Pomerium v0.22, the JWT issuer claim will be set to the route
domain, rather than the authenticate service domain. (That is, it will
have the same value as the subject claim.) Update the JS SDK guide page
to reflect this change.

Update the JWT payload example output to use a code block rather than a
screenshot, to make this easier to modify.