start using 'shell-session' code blocks
Configure 'shell-session' as an additional language for syntax
highlighting in code blocks. This differs from 'bash' highlighting in
that lines not beginning with a shell symbol (e.g. '$') won't be treated
as part of a shell command.

Some of our existing code blocks use 'bash' along with line highlights
to present interactive shell sessions. Convert these to use
'shell-session', adding '$' to any input lines.

Note that the prism-react-renderer themes do not currently have styles
for the 'shell-session' token types, so we need to define our own styles
for these.