incorporate feedback, add missing install step
Update keyword metadata, remove version number mention, and add another
sentence to the overview section.

Add a missing installation step: creating a global configuration object.
Add another policy example in the section introducing the PolicyFilter
CRD. Add an admonition to this section to emphasize the importance of
adding a PolicyFilter to each route. Tweak some of the wording.