rework Host header settings reference
Start by spelling out the default behavior if no host header settings
are provided. Re-order the page to put the description of the different
settings first, before the table of config keys. Make sure to include
preserve_host_header in the table.

Add a note that these settings apply just the same to the :authority
pseudo-header (for HTTP/2 and HTTP/3 requests).