add reference page for depends_on feature (#1861)
Summarize the behavior and use case for the depends_on route option. 
Spell out the current limitations of the feature. Include a screenshot
of the Enterprise Console UI.

Add the new Ingress annotation to the list of Pomerium-standard
annotations on the k8s docs page.

Update the troubleshooting section on CORS issues with a section on this
feature and add links to and from the reference page.