pre-commit: update, clarify dev setup, and cleanup
We use the python pre-commit system in this repo with a separate github
action for it, but the configuration was outdated and I'm not sure any
developers were setting it up locally, so this clarifies the
installation requirements and updates related configs.

I also removed the cspell npm scripts, because you don't run them with
yarn/npm.